fix: resolve all 35 SonarCloud issues in phase-2 PR
Apply C++20/23 modernization and static-analysis fixes across 11 files:
- S995: const-qualify pointer params (mem_lock/mem_unlock) with Windows const_cast
- S1231: replace malloc/free with new[]/delete[] in SecureBuffer
- S6494: replace fprintf with std::println (C++23)
- S6004: move variable declarations into if-init statements
- S7035: replace static_cast<uint8_t>(enum) with std::to_underlying
- S5945: replace C-style arrays with std::array (MAGIC, IndexSlot slot[2])
- S125: remove commented-out dead code
- S5276: add outer explicit cast for implicit int narrowing
- S1117: rename off::MAGIC → off::MAGIC_OFF to avoid shadowing vault::MAGIC
- S5566: replace range-for with std::ranges::any_of
- S3574: remove redundant lambda return type annotation
- S6177: use enum VaultResult in all vault.cpp functions
- S6229: replace std::time(nullptr) with std::chrono::system_clock
- S1068: mark Phase 0 placeholder field [[maybe_unused]]

+++ src/vault/byte_io.h
@@ -45,8 +45,9 @@ inline void put_bytes_at(std::span<uint8_t> buf, size_t off,
 
 inline uint16_t get_u16_at(std::span<const uint8_t> buf, size_t off) noexcept
 {
-    return static_cast<uint16_t>(buf[off + 0]) |
-           static_cast<uint16_t>(buf[off + 1]) << 8;
+    return static_cast<uint16_t>(
+        static_cast<unsigned>(buf[off + 0]) |
+        (static_cast<unsigned>(buf[off + 1]) << 8));
 }
 
 inline uint32_t get_u32_at(std::span<const uint8_t> buf, size_t off) noexcept
